Split up converter
