+++ packages/data-sql/src/converter/index.ts
@@ -1,5 +1,6 @@
-import type { AbstractQuery, AbstractQueryFieldNodePrimitive } from '@directus/data/types';
-import type { SqlStatement, SqlStatementSelectPrimitive } from './types.js';
+import type { AbstractQuery } from '@directus/data/types';
+import type { SqlStatement } from '../types.js';
+import { convertPrimitive } from './convert-primitive.js';
 
 /**
  * @param abstractQuery the abstract query to convert
@@ -22,28 +23,6 @@ export const convertAbstractQueryToSqlStatement = (abstractQuery: AbstractQuery)
 		}),
 		from: abstractQuery.collection,
 	};
-
-	return statement;
-};
-
-/**
- * @param abstractPrimitive
- * @param collection
- * @returns the converted primitive node
- */
-export const convertPrimitive = (
-	abstractPrimitive: AbstractQueryFieldNodePrimitive,
-	collection: string
-): SqlStatementSelectPrimitive => {
-	const statement: SqlStatementSelectPrimitive = {
-		type: 'primitive',
-		table: collection,
-		column: abstractPrimitive.field,
-	};
-
-	if (abstractPrimitive.alias) {
-		statement.as = abstractPrimitive.alias;
-	}
 
 	return statement;
 };
